refactor: common `TanstackTable` component
