+++ frontend/src/components/Table/TableEmpty.vue
